perf(display): 优化 GIF播放性能
- 将 Config 结构体中的 fps 字段类型从 usize 改为 u64
- 优化 GIF 播放的延迟计算和处理逻辑，提高帧率精度
-调整 sleep 延迟，考虑渲染时间，避免累积误差

+++ src/config.rs
@@ -137,7 +137,7 @@ pub struct DirectoryArgs {
 #[derive(Parser)]
 pub struct GifArgs {
     #[clap(long, help = "Set the frames per second for gif playback", default_value = "10")]
-    pub fps: usize,
+    pub fps: u64,
     #[clap(long = "loop", help = "Loop the gif playback", default_value_t = false)]
     pub loop_play: bool,
     #[clap(help = "Gif file path")]
@@ -185,7 +185,7 @@ impl Default for Cli {
 
 #[derive(Debug, Clone)]
 pub struct Config {
-    pub fps: usize,
+    pub fps: u64,
     pub pause: bool,
     pub center: bool,
     pub no_color: bool,
